requirements.txt
pytest==8.3.5
selenium==4.31.0
allure-pytest==2.14.0
